requirements.txt
amqp==5.3.1
asgiref==3.8.1
async-timeout==5.0.1
billiard==4.2.1
celery==5.5.2
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cron-descriptor==1.4.5
diff-match-patch==20241021
Django==5.1.7
django-celery-beat==2.8.1
django-import-export==4.3.7
django-timezone-field==7.1
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
dotenv==0.9.9
Faker==37.3.0
gunicorn==23.0.0
kombu==5.5.3
loadenv==0.1.1
packaging==25.0
pillow==11.1.0
prompt_toolkit==3.0.51
psycopg==3.2.9
psycopg2-binary==2.9.10
PyJWT==2.9.0
python-crontab==3.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
redis==6.1.0
six==1.17.0
sqlparse==0.5.3
tablib==3.8.0
typing_extensions==4.13.2
tzdata==2025.2
uuid==1.30
vine==5.1.0
wcwidth==0.2.13
